chore: 팀스페이스 생성/입장 파일 분리
